[fix] #93 repository ã상속 시 제네릭 타입 불일치 수정

+++ module-domain/domain-inventories/src/main/java/com/chaing/domain/inventories/repository/FactoryInventoryRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FactoryInventoryRepository extends JpaRepository<FactoryInventory,Integer> {
+public interface FactoryInventoryRepository extends JpaRepository<FactoryInventory,Long> {
 
     List<FactoryInventory> findAllByStatus(LogType status);
